feat: configure system services

+++ internal/algod/linux/linux.go
@@ -1,7 +1,7 @@
 package linux
 
 import (
-	"bytes"
+	"errors"
 	"fmt"
 	"github.com/algorandfoundation/nodekit/internal/algod/fallback"
 	"github.com/algorandfoundation/nodekit/internal/system"
@@ -10,11 +10,12 @@ import (
 	"os/exec"
 	"runtime"
 	"strings"
-	"text/template"
 )
 
 // PackageManagerNotFoundMsg is an error message indicating the absence of a supported package manager for uninstalling Algorand.
 const PackageManagerNotFoundMsg = "could not find a package manager to uninstall Algorand"
+const SystemdOverridesDir = "/etc/systemd/system/algorand.service.d"
+const SystemdOverrideFilePath = SystemdOverridesDir + "/override.conf"
 
 // Algod represents an implementation of the system.Interface tailored for managing the Algod service.
 // It includes details about the service's executable path and associated data directory.
@@ -46,12 +47,12 @@ func InstallRequirements() system.CmdsList {
 }
 
 // Install installs Algorand development tools or node software depending on the package manager.
-func Install() error {
+func Install(dataDir string, network string, force bool) error {
 	log.Info("Installing Algod on Linux")
 	// Based off of https://developer.algorand.org/docs/run-a-node/setup/install/#installation-with-a-package-manager
 	if system.CmdExists("apt-get") { // On some Debian systems we use apt-get
 		log.Info("Installing with apt-get")
-		return system.RunAll(append(InstallRequirements(), system.CmdsList{
+		err := system.RunAll(append(InstallRequirements(), system.CmdsList{
 			{"sudo", "apt-get", "update"},
 			{"sudo", "apt-get", "install", "-y", "gnupg2", "curl", "software-properties-common"},
 			{"sh", "-c", "curl -o - https://releases.algorand.com/key.pub | sudo tee /etc/apt/trusted.gpg.d/algorand.asc"},
@@ -59,11 +60,14 @@ func Install() error {
 			{"sudo", "apt-get", "update"},
 			{"sudo", "apt-get", "install", "-y", "algorand"},
 		}...))
+		if err != nil {
+			return err
+		}
 	}
 
 	if system.CmdExists("dnf") { // On Fedora and CentOs8 there's the dnf package manager
 		log.Printf("Installing with dnf")
-		return system.RunAll(append(InstallRequirements(), system.CmdsList{
+		err := system.RunAll(append(InstallRequirements(), system.CmdsList{
 			{"curl", "-O", "https://releases.algorand.com/rpm/rpm_algorand.pub"},
 			{"sudo", "rpmkeys", "--import", "rpm_algorand.pub"},
 			{"sudo", "dnf", "install", "-y", "dnf-command(config-manager)"},
@@ -73,42 +77,48 @@ func Install() error {
 			{"sudo", "systemctl", "start", "algorand.service"},
 			{"rm", "-f", "rpm_algorand.pub"},
 		}...))
-
+		if err != nil {
+			return err
+		}
 	}
 
-	// TODO: watch this method to see if it is ever used
-	return fallback.Install()
+	return fallback.SetNetwork(dataDir, network, force)
 }
 
 // Uninstall removes the Algorand software using a supported package manager or clears related system files if necessary.
 // Returns an error if a supported package manager is not found or if any command fails during execution.
-func Uninstall() error {
-	log.Info("Uninstalling Algorand")
-	var unInstallCmds system.CmdsList
-	// On Ubuntu and Debian there's the apt package manager
-	if system.CmdExists("apt-get") {
-		log.Info("Using apt-get package manager")
-		unInstallCmds = [][]string{
-			{"sudo", "apt-get", "autoremove", "algorand", "-y"},
+func Uninstall(dataDir string, force bool) error {
+	if dataDir != "" {
+		log.Info("Uninstalling Algorand")
+		var unInstallCmds system.CmdsList
+		// On Ubuntu and Debian there's the apt package manager
+		if system.CmdExists("apt-get") {
+			log.Info("Using apt-get package manager")
+			unInstallCmds = [][]string{
+				{"sudo", "apt-get", "autoremove", "algorand", "-y"},
+			}
 		}
-	}
-	// On Fedora and CentOs8 there's the dnf package manager
-	if system.CmdExists("dnf") {
-		log.Info("Using dnf package manager")
-		unInstallCmds = [][]string{
-			{"sudo", "dnf", "remove", "algorand", "-y"},
+		// On Fedora and CentOs8 there's the dnf package manager
+		if system.CmdExists("dnf") {
+			log.Info("Using dnf package manager")
+			unInstallCmds = [][]string{
+				{"sudo", "dnf", "remove", "algorand", "-y"},
+			}
 		}
-	}
-	// Error on unsupported package managers
-	if len(unInstallCmds) == 0 {
-		return fmt.Errorf(PackageManagerNotFoundMsg)
-	}
+		// Error on unsupported package managers
+		if len(unInstallCmds) == 0 {
+			return fmt.Errorf(PackageManagerNotFoundMsg)
+		}
 
-	// Commands to clear systemd algorand.service and any other files, like the configuration override
-	unInstallCmds = append(unInstallCmds, []string{"sudo", "bash", "-c", "rm -rf /etc/systemd/system/algorand*"})
-	unInstallCmds = append(unInstallCmds, []string{"sudo", "systemctl", "daemon-reload"})
+		// Commands to clear systemd algorand.service and any other files, like the configuration override
+		unInstallCmds = append(unInstallCmds, []string{"sudo", "bash", "-c", "rm -rf /etc/systemd/system/algorand*"})
+		unInstallCmds = append(unInstallCmds, []string{"sudo", "systemctl", "daemon-reload"})
 
-	return system.RunAll(unInstallCmds)
+		return system.RunAll(unInstallCmds)
+	}
+	// TODO: fallback uninstall
+	//return fallback.Uninstall(dataDir)
+	return errors.New("fallback uninstall not supported on Linux")
 }
 
 // Upgrade updates Algorand and its dev tools using an approved package
@@ -132,92 +142,33 @@ func Upgrade() error {
 // It executes the appropriate command for systemd on Linux-based systems.
 // Returns an error if the command fails.
 // TODO: Replace with D-Bus integration
-func Start() error {
+func Start(dataDir string) error {
+	// Just start the service when no directory is specified
+	//if dataDir == "" {
 	return exec.Command("sudo", "systemctl", "start", "algorand").Run()
+	//}
+	//return fallback.Start(dataDir)
 }
 
 // Stop shuts down the Algorand algod system process on Linux using the systemctl stop command.
 // Returns an error if the operation fails.
 // TODO: Replace with D-Bus integration
-func Stop() error {
-	return exec.Command("sudo", "systemctl", "stop", "algorand").Run()
+func Stop(dataDir string) error {
+	// Just stop the service when no directory is specified
+	if dataDir == "" {
+		return exec.Command("sudo", "systemctl", "stop", "algorand").Run()
+	}
+	// TODO: Deprecate the forceful stopping
+	return fallback.Stop(dataDir)
 }
 
 // IsService checks if the "algorand.service" is listed as a systemd unit file on Linux.
 // Returns true if it exists.
 // TODO: Replace with D-Bus integration
-func IsService() bool {
-	out, err := system.Run([]string{"sudo", "systemctl", "list-unit-files", "algorand.service"})
+func IsService(dataDir string) bool {
+	out, err := system.Run([]string{"sudo", "systemctl", "list-unit-files", "algorand*"})
 	if err != nil {
 		return false
 	}
 	return strings.Contains(out, "algorand.service")
-}
-
-// UpdateService updates the systemd service file for the Algorand daemon
-// with a new data directory path and reloads the daemon.
-func UpdateService(dataDirectoryPath string) error {
-
-	algodPath, err := exec.LookPath("algod")
-	if err != nil {
-		fmt.Printf("Failed to find algod binary: %v\n", err)
-		os.Exit(1)
-	}
-
-	// Path to the systemd service override file
-	// Assuming that this is the same everywhere systemd is used
-	overrideFilePath := "/etc/systemd/system/algorand.service.d/override.conf"
-
-	// Create the override directory if it doesn't exist
-	err = os.MkdirAll("/etc/systemd/system/algorand.service.d", 0755)
-	if err != nil {
-		fmt.Printf("Failed to create override directory: %v\n", err)
-		os.Exit(1)
-	}
-
-	// Content of the override file
-	const overrideTemplate = `[Unit]
-Description=Algorand daemon {{.AlgodPath}} in {{.DataDirectoryPath}}
-[Service]
-ExecStart=
-ExecStart={{.AlgodPath}} -d {{.DataDirectoryPath}}`
-
-	// Data to fill the template
-	data := map[string]string{
-		"AlgodPath":         algodPath,
-		"DataDirectoryPath": dataDirectoryPath,
-	}
-
-	// Parse and execute the template
-	tmpl, err := template.New("override").Parse(overrideTemplate)
-	if err != nil {
-		fmt.Printf("Failed to parse template: %v\n", err)
-		os.Exit(1)
-	}
-
-	var overrideContent bytes.Buffer
-	err = tmpl.Execute(&overrideContent, data)
-	if err != nil {
-		fmt.Printf("Failed to execute template: %v\n", err)
-		os.Exit(1)
-	}
-
-	// Write the override content to the file
-	err = os.WriteFile(overrideFilePath, overrideContent.Bytes(), 0644)
-	if err != nil {
-		fmt.Printf("Failed to write override file: %v\n", err)
-		os.Exit(1)
-	}
-
-	// Reload systemd manager configuration
-	cmd := exec.Command("systemctl", "daemon-reload")
-	err = cmd.Run()
-	if err != nil {
-		fmt.Printf("Failed to reload systemd daemon: %v\n", err)
-		os.Exit(1)
-	}
-
-	log.Info("Algorand service file updated successfully.")
-
-	return nil
 }
